moved calibration unit test to own file
